server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
requests>=2.28.0
